final product

+++ ft_calloc.c
@@ -6,7 +6,7 @@
 /*   By: atereso- <atereso-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/28 16:17:04 by atereso-          #+#    #+#             */
-/*   Updated: 2021/11/25 16:57:06 by atereso-         ###   ########.fr       */
+/*   Updated: 2021/12/21 13:17:07 by atereso-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_calloc(size_t count, size_t size)
 {
 	char	*allocator;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	allocator = malloc(count * size);
